feat(core): product requirement changes. remove `mode` field, the mode depends on whether there is a filter
