+++ packages/frontend/component/src/components/page-list/view/index.ts
@@ -3,5 +3,5 @@ export * from './collection-bar';
 export * from './collection-list';
 export * from './collection-operations';
 export * from './create-collection';
-export * from './edit-collection';
+export * from './edit-collection/edit-collection';
 export * from './use-edit-collection';
