+++ packages/frontend/i18n/src/resources/en.json
@@ -407,16 +407,19 @@
   "com.affine.editCollectionName.name": "Name",
   "com.affine.editCollectionName.name.placeholder": "Collection Name",
   "com.affine.editCollectionName.createTips": "Collection is a smart folder where you can manually add pages or automatically add pages through rules.",
+  "com.affine.editCollection.rules.include.title": "Selected pages",
   "com.affine.editCollection.rules.include.page": "Page",
   "com.affine.editCollection.rules.include.is": "is",
-  "com.affine.editCollection.rules.include.add": "Add include page",
-  "com.affine.editCollection.rules.include.tipsTitle": "What is \"Include\"？",
-  "com.affine.editCollection.rules.include.tips": "\"Include\" refers to manually adding pages rather than automatically adding them through rule matching. You can manually add pages through the \"Add pages\" option or by dragging and dropping (coming soon).",
+  "com.affine.editCollection.rules.include.add": "Add selected page",
+  "com.affine.editCollection.rules.include.tipsTitle": "What is \"Selected pages\"？",
+  "com.affine.editCollection.rules.include.tips": "“Selected pages” refers to manually adding pages rather than automatically adding them through rule matching. You can manually add pages through the “Add selected pages” option or by dragging and dropping.",
   "com.affine.editCollection.rules.preview": "Preview",
   "com.affine.editCollection.rules.reset": "Reset",
-  "com.affine.editCollection.rules.countTips.zero": "Showing <1>{{count}}</1> pages.",
-  "com.affine.editCollection.rules.countTips.one": "Showing <1>{{count}}</1> page.",
-  "com.affine.editCollection.rules.countTips.more": "Showing <1>{{count}}</1> pages.",
+  "com.affine.editCollection.rules.countTips": "Selected <1>{{selectedCount}}</1>, filtered <3>{{filteredCount}}</3>",
+  "com.affine.editCollection.rules.empty.noRules": "No Rules",
+  "com.affine.editCollection.rules.empty.noRules.tips": "Please <1>add rules</1> to save this collection or switch to <3>Pages</3>, use manual selection mode",
+  "com.affine.editCollection.rules.empty.noResults": "No Results",
+  "com.affine.editCollection.rules.empty.noResults.tips": "No pages meet the filtering rules",
   "com.affine.selectPage.title": "Add include page",
   "com.affine.selectPage.selected": "Selected",
   "com.affine.selectPage.empty": "Empty",
